Change:Like-model

+++ back/app/models/like.model.ts
@@ -1,5 +1,5 @@
 import { pool } from './db';
-import { ResultSetHeader, RowDataPacket } from 'mysql2';
+import { RowDataPacket,  FieldPacket} from 'mysql2';
 
 /*export interface Like{
     recipe_id: Number;
@@ -18,55 +18,49 @@ export class Likemodel {
         this.user_id = like.user_id;
     }
     
+//Añadir like
     static async addLike(recipeId: Number, userId: Number): Promise<void> {
-        try{
-            const [existingLikes] = await this.connection.query(
-                'SELECT * FROM likes WHERE recipe_id = ? AND user_id = ?',
-                [recipeId, userId]
-            );
+      const connection = await pool.getConnection();
+      try {
+        const [existingLikes] = await connection.query('SELECT * FROM likes WHERE recipe_id = ? AND user_id = ?', [recipeId, userId]);
 
-            if(existingLikes.length === 0) {
-               await this.connection.query(
-                'INSERT INTO likes (recipi_id, user_id) VALUES (?, ?)',
-                [recipeId, userId]
-               );
+        if ((existingLikes as any[]).length === 0) {
+            await connection.query('INSERT INTO likes (recipe_id, user_id) VALUES (?, ?)', [recipeId, userId]);
             console.log('Like added successfully.');
-            } else {
-                console.log('The user has already "Liked" this recipe.');
-            } catch (error) {
-                console.error('Error adding "Like":', error);
-                throw error;
-            }
+        } else {
+            console.log('The user has already "Liked" this recipe.');
         }
+    } catch (error) {
+        console.error('Error adding "Like":', error);
+        throw error;
     }
-
+}   
+   
+//Eliminar like
     static async removeLike(recipeId: Number, userId: Number): Promise<void> {
+      const connection = await pool.getConnection();
         try {
-            const [existingLikes] = await this.connection.query(
-                'SELECT * FROM likes WHERE recipe_id = ? AND user_id = ?',
-                [recipeId, userId]
-            );
-        if (existingLikes > 0){
-            await this.connection.query(
-                'DELETE FROM likes WHERE recipe_id = ? AND user_id = ?',
-                [recipeId, userId]
-            );
-            console.log('Like successfully removed.');
-        }else{
+          //  const [existingLikes]: [RowDataPacket[]] = await connection.query('SELECT * FROM likes WHERE recipe_id = ? AND user_id = ?',[recipeId, userId]);
+            const [existingLikes, _]: [RowDataPacket[], FieldPacket[]] = await connection.query('SELECT * FROM likes WHERE recipe_id = ? AND user_id = ?', [recipeId, userId]); 
+            if (existingLikes.length > 0) {
+                await connection.query('DELETE FROM likes WHERE recipe_id = ? AND user_id = ?',[recipeId, userId]);
+                console.log('Like successfully removed.');
+         } else {
             console.log('The user has not "Liked" this recipe.');
-        } catch (Error){
+        } 
+        }catch (error){
             console.error('Error deleting likes:', error);
             throw error;
         }
-      }
     }
 
-    static async getLikesCount(recipeId): Promise<Number>{
+    //Recuento de Like
+    static async getLikesCount(recipeId: Number): Promise<Number>{
+        const connection = await pool.getConnection();
         try{
-           const [rows] = await this.connection.query(
+           const [rows]: [RowDataPacket[], FieldPacket[]] = await connection.query(
             'SELECT COUNT(*) as count FROM likes WHERE recipe_id = ?',
-            [recipeId]
-          );
+            [recipeId]);
            
           const likesCount: number = rows[0].count;
